Fixing cost issue in the corridor rasters
The corridor rasters didn't have 0 values on the least-cost paths between the patches, as should have been. This was due to a mismatch between the computation of the least-cost path (done with diagonals) and the CWD raster by whitebox (done with diagonals, but also with multiplication by cell size). Everything should now be fixed.